backend/requirements.txt
Django==4.2
django-cors-headers==4.2
python-dotenv==1.0
google-generativeai==0.3.0
tavily-python==0.1.0